Fix: Promises not discarded with async validation
Upon asynchronous form validation, the current pending validation
promise was not discarded when a new validation promise is created. This
could lead to an incorrect validation state when the first promise is
fulfilled after the new one.

The test added with this commit also tests that the recently added
debounced asynchronous validation works.

Also: In `DropDownField`, expect the "value" and options "value" props
to be of type `string` with prop types validation. This is necessary
after the dropdown value handling change in `react-components` v0.2.3,
and is also compliant with the "value" attribute of HTML `<option>`s.